pdf_service/requirements.txt
Flask==3.0.3
confluent-kafka==2.4.0
python-dotenv==1.0.1
requests==2.32.3